feat: allow checking multiple ports
I actually made two sets of these buttons for work and home. So what
this does is it checks all ports listed to see if the buttons work. In
the future, I think it would be better for the buttons firmware to have
a simple hello command to test if the port is actually buttons. Then
it can just try all ports and use that hello command.